chore: restore test runner
