+++ java-engine/src/test/java/io/getunleash/engine/LibraryTest.java
@@ -3,7 +3,7 @@
  */
 package io.getunleash.engine;
 
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class LibraryTest {
   @Test
